+++ core/src/avm1/globals.rs
@@ -1,10 +1,10 @@
 use crate::avm1::fscommand;
 use crate::avm1::function::Executable;
 use crate::avm1::return_value::ReturnValue;
-use crate::avm1::{Avm1, Error, Object, UpdateContext, Value};
+use crate::avm1::{Avm1, Error, Object, ObjectCell, ScriptObject, UpdateContext, Value};
 use crate::backend::navigator::NavigationMethod;
 use enumset::EnumSet;
-use gc_arena::{GcCell, MutationContext};
+use gc_arena::MutationContext;
 use rand::Rng;
 
 mod function;
@@ -16,7 +16,7 @@ mod object;
 pub fn getURL<'a, 'gc>(
     avm: &mut Avm1<'gc>,
     context: &mut UpdateContext<'a, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     //TODO: Error behavior if no arguments are present
@@ -44,7 +44,7 @@ pub fn getURL<'a, 'gc>(
 pub fn random<'gc>(
     _avm: &mut Avm1<'gc>,
     action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     match args.get(0) {
@@ -56,7 +56,7 @@ pub fn random<'gc>(
 pub fn boolean<'gc>(
     avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     if let Some(val) = args.get(0) {
@@ -69,7 +69,7 @@ pub fn boolean<'gc>(
 pub fn number<'gc>(
     _avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     if let Some(val) = args.get(0) {
@@ -82,7 +82,7 @@ pub fn number<'gc>(
 pub fn is_nan<'gc>(
     _avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     if let Some(val) = args.get(0) {
@@ -97,9 +97,9 @@ pub fn is_nan<'gc>(
 /// user-modifiable.
 #[derive(Clone)]
 pub struct SystemPrototypes<'gc> {
-    pub object: GcCell<'gc, Object<'gc>>,
-    pub function: GcCell<'gc, Object<'gc>>,
-    pub movie_clip: GcCell<'gc, Object<'gc>>,
+    pub object: ObjectCell<'gc>,
+    pub function: ObjectCell<'gc>,
+    pub movie_clip: ObjectCell<'gc>,
 }
 
 unsafe impl<'gc> gc_arena::Collect for SystemPrototypes<'gc> {
@@ -114,36 +114,37 @@ unsafe impl<'gc> gc_arena::Collect for SystemPrototypes<'gc> {
 /// Initialize default global scope and builtins for an AVM1 instance.
 pub fn create_globals<'gc>(
     gc_context: MutationContext<'gc, '_>,
-) -> (SystemPrototypes<'gc>, Object<'gc>) {
-    let object_proto = GcCell::allocate(gc_context, Object::bare_object());
+) -> (SystemPrototypes<'gc>, Box<dyn Object<'gc> + 'gc>) {
+    let object_proto = ScriptObject::object_cell(gc_context, None);
     let function_proto = function::create_proto(gc_context, object_proto);
 
     object::fill_proto(gc_context, object_proto, function_proto);
 
-    let movie_clip_proto = movie_clip::create_proto(gc_context, object_proto, function_proto);
+    let movie_clip_proto: ObjectCell<'gc> =
+        movie_clip::create_proto(gc_context, object_proto, function_proto);
 
     //TODO: These need to be constructors and should also set `.prototype` on each one
-    let object = Object::function(
+    let object = ScriptObject::function(
         gc_context,
         Executable::Native(object::constructor),
         Some(function_proto),
         Some(object_proto),
     );
 
-    let function = Object::function(
+    let function = ScriptObject::function(
         gc_context,
         Executable::Native(function::constructor),
         Some(function_proto),
         Some(function_proto),
     );
-    let movie_clip = Object::function(
+    let movie_clip = ScriptObject::function(
         gc_context,
         Executable::Native(movie_clip::constructor),
         Some(function_proto),
         Some(movie_clip_proto),
     );
 
-    let mut globals = Object::object(gc_context, Some(object_proto));
+    let mut globals = ScriptObject::object(gc_context, Some(object_proto));
     globals.force_set("Object", object, EnumSet::empty());
     globals.force_set("Function", function, EnumSet::empty());
     globals.force_set("MovieClip", movie_clip, EnumSet::empty());
@@ -204,7 +205,7 @@ pub fn create_globals<'gc>(
             function: function_proto,
             movie_clip: movie_clip_proto,
         },
-        globals,
+        Box::new(globals),
     )
 }
