+++ core/src/avm1/globals/movie_clip.rs
@@ -1,10 +1,9 @@
 //! MovieClip prototype
 
 use crate::avm1::function::Executable;
-use crate::avm1::object::Object;
 use crate::avm1::property::Attribute::*;
 use crate::avm1::return_value::ReturnValue;
-use crate::avm1::{Avm1, Error, UpdateContext, Value};
+use crate::avm1::{Avm1, Error, ObjectCell, ScriptObject, UpdateContext, Value};
 use crate::display_object::{DisplayNode, DisplayObject, MovieClip};
 use enumset::EnumSet;
 use gc_arena::{GcCell, MutationContext};
@@ -13,7 +12,7 @@ use gc_arena::{GcCell, MutationContext};
 pub fn constructor<'gc>(
     _avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     _args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     Ok(Value::Undefined.into())
@@ -25,7 +24,7 @@ macro_rules! with_movie_clip {
             $object.force_set_function(
                 $name,
                 |_avm, _context, this, args| -> Result<ReturnValue<'gc>, Error> {
-                    if let Some(display_object) = this.read().display_node() {
+                    if let Some(display_object) = this.read().as_display_node() {
                         if let Some(movie_clip) = display_object.read().as_movie_clip() {
                             return Ok($fn(movie_clip, args));
                         }
@@ -46,7 +45,7 @@ macro_rules! with_movie_clip_mut {
             $object.force_set_function(
                 $name,
                 |_avm, context: &mut UpdateContext<'_, 'gc, '_>, this, args| -> Result<ReturnValue<'gc>, Error> {
-                    if let Some(display_object) = this.read().display_node() {
+                    if let Some(display_object) = this.read().as_display_node() {
                         if let Some(movie_clip) = display_object.write(context.gc_context).as_movie_clip_mut() {
                             return Ok($fn(movie_clip, context, display_object, args).into());
                         }
@@ -64,7 +63,7 @@ macro_rules! with_movie_clip_mut {
 pub fn overwrite_root<'gc>(
     _avm: &mut Avm1<'gc>,
     ac: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     let new_val = args
@@ -72,6 +71,8 @@ pub fn overwrite_root<'gc>(
         .map(|v| v.to_owned())
         .unwrap_or(Value::Undefined);
     this.write(ac.gc_context)
+        .as_script_object_mut()
+        .unwrap()
         .force_set("_root", new_val, EnumSet::new());
 
     Ok(Value::Undefined.into())
@@ -80,7 +81,7 @@ pub fn overwrite_root<'gc>(
 pub fn overwrite_global<'gc>(
     _avm: &mut Avm1<'gc>,
     ac: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     let new_val = args
@@ -88,6 +89,8 @@ pub fn overwrite_global<'gc>(
         .map(|v| v.to_owned())
         .unwrap_or(Value::Undefined);
     this.write(ac.gc_context)
+        .as_script_object_mut()
+        .unwrap()
         .force_set("_global", new_val, EnumSet::new());
 
     Ok(Value::Undefined.into())
@@ -95,10 +98,10 @@ pub fn overwrite_global<'gc>(
 
 pub fn create_proto<'gc>(
     gc_context: MutationContext<'gc, '_>,
-    proto: GcCell<'gc, Object<'gc>>,
-    fn_proto: GcCell<'gc, Object<'gc>>,
-) -> GcCell<'gc, Object<'gc>> {
-    let mut object = Object::object(gc_context, Some(proto));
+    proto: ObjectCell<'gc>,
+    fn_proto: ObjectCell<'gc>,
+) -> ObjectCell<'gc> {
+    let mut object = ScriptObject::object(gc_context, Some(proto));
 
     with_movie_clip_mut!(
         gc_context,
@@ -158,7 +161,7 @@ pub fn create_proto<'gc>(
         Executable::Native(|_avm, _context, this, _args| {
             Ok(this
                 .read()
-                .display_node()
+                .as_display_node()
                 .and_then(|mc| mc.read().parent())
                 .and_then(|dn| dn.read().object().as_object().ok())
                 .map(|o| Value::Object(o.to_owned()))
@@ -169,5 +172,5 @@ pub fn create_proto<'gc>(
         EnumSet::new(),
     );
 
-    GcCell::allocate(gc_context, object)
+    GcCell::allocate(gc_context, Box::new(object))
 }
