+++ core/src/avm1/globals/object.rs
@@ -1,8 +1,7 @@
 //! Object prototype
 use crate::avm1::property::Attribute::*;
 use crate::avm1::return_value::ReturnValue;
-use crate::avm1::{Avm1, Error, Object, Value};
-use crate::context::UpdateContext;
+use crate::avm1::{Avm1, Error, ObjectCell, UpdateContext, Value};
 use enumset::EnumSet;
 use gc_arena::{GcCell, MutationContext};
 
@@ -10,7 +9,7 @@ use gc_arena::{GcCell, MutationContext};
 pub fn constructor<'gc>(
     _avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: GcCell<'gc, Object<'gc>>,
+    _this: ObjectCell<'gc>,
     _args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     Ok(Value::Undefined.into())
@@ -20,7 +19,7 @@ pub fn constructor<'gc>(
 pub fn add_property<'gc>(
     _avm: &mut Avm1<'gc>,
     context: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     let name = args.get(0).unwrap_or(&Value::Undefined);
@@ -32,17 +31,17 @@ pub fn add_property<'gc>(
             if let Some(get_func) = get.read().as_executable() {
                 if let Value::Object(set) = setter {
                     if let Some(set_func) = set.read().as_executable() {
-                        this.write(context.gc_context).force_set_virtual(
-                            name,
-                            get_func,
-                            Some(set_func),
-                            EnumSet::empty(),
-                        );
+                        this.write(context.gc_context)
+                            .as_script_object_mut()
+                            .unwrap()
+                            .force_set_virtual(name, get_func, Some(set_func), EnumSet::empty());
                     } else {
                         return Ok(false.into());
                     }
                 } else if let Value::Null = setter {
                     this.write(context.gc_context)
+                        .as_script_object_mut()
+                        .unwrap()
                         .force_set_virtual(name, get_func, None, ReadOnly);
                 } else {
                     return Ok(false.into());
@@ -59,7 +58,7 @@ pub fn add_property<'gc>(
 pub fn has_own_property<'gc>(
     _avm: &mut Avm1<'gc>,
     _action_context: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     match args.get(0) {
@@ -72,7 +71,7 @@ pub fn has_own_property<'gc>(
 fn to_string<'gc>(
     _: &mut Avm1<'gc>,
     _: &mut UpdateContext<'_, 'gc, '_>,
-    _: GcCell<'gc, Object<'gc>>,
+    _: ObjectCell<'gc>,
     _: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     Ok(ReturnValue::Immediate("[Object object]".into()))
@@ -82,7 +81,7 @@ fn to_string<'gc>(
 fn is_property_enumerable<'gc>(
     _: &mut Avm1<'gc>,
     _: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     match args.get(0) {
@@ -97,19 +96,28 @@ fn is_property_enumerable<'gc>(
 fn is_prototype_of<'gc>(
     _: &mut Avm1<'gc>,
     _: &mut UpdateContext<'_, 'gc, '_>,
-    this: GcCell<'gc, Object<'gc>>,
+    this: ObjectCell<'gc>,
     args: &[Value<'gc>],
 ) -> Result<ReturnValue<'gc>, Error> {
     match args.get(0) {
-        Some(Value::Object(ob)) => {
-            let mut proto = ob.read().prototype().cloned();
+        Some(val) => {
+            let ob = match val.as_object() {
+                Ok(ob) => ob,
+                Err(_) => return Ok(Value::Bool(false).into()),
+            };
+            let mut proto = ob.read().as_script_object().unwrap().prototype().cloned();
 
             while let Some(proto_ob) = proto {
                 if GcCell::ptr_eq(this, proto_ob) {
                     return Ok(Value::Bool(true).into());
                 }
 
-                proto = proto_ob.read().prototype().cloned();
+                proto = proto_ob
+                    .read()
+                    .as_script_object()
+                    .unwrap()
+                    .prototype()
+                    .cloned();
             }
 
             Ok(Value::Bool(false).into())
@@ -129,44 +137,59 @@ fn is_prototype_of<'gc>(
 /// bare objects for both and let this function fill Object for you.
 pub fn fill_proto<'gc>(
     gc_context: MutationContext<'gc, '_>,
-    object_proto: GcCell<'gc, Object<'gc>>,
-    fn_proto: GcCell<'gc, Object<'gc>>,
+    object_proto: ObjectCell<'gc>,
+    fn_proto: ObjectCell<'gc>,
 ) {
     let mut ob_proto_write = object_proto.write(gc_context);
 
-    ob_proto_write.force_set_function(
-        "addProperty",
-        add_property,
-        gc_context,
-        DontDelete | DontEnum,
-        Some(fn_proto),
-    );
-    ob_proto_write.force_set_function(
-        "hasOwnProperty",
-        has_own_property,
-        gc_context,
-        DontDelete | DontEnum,
-        Some(fn_proto),
-    );
-    ob_proto_write.force_set_function(
-        "isPropertyEnumerable",
-        is_property_enumerable,
-        gc_context,
-        DontDelete | DontEnum,
-        Some(fn_proto),
-    );
-    ob_proto_write.force_set_function(
-        "isPrototypeOf",
-        is_prototype_of,
-        gc_context,
-        DontDelete | DontEnum,
-        Some(fn_proto),
-    );
-    ob_proto_write.force_set_function(
-        "toString",
-        to_string,
-        gc_context,
-        DontDelete | DontEnum,
-        Some(fn_proto),
-    );
+    ob_proto_write
+        .as_script_object_mut()
+        .unwrap()
+        .force_set_function(
+            "addProperty",
+            add_property,
+            gc_context,
+            DontDelete | DontEnum,
+            Some(fn_proto),
+        );
+    ob_proto_write
+        .as_script_object_mut()
+        .unwrap()
+        .force_set_function(
+            "hasOwnProperty",
+            has_own_property,
+            gc_context,
+            DontDelete | DontEnum,
+            Some(fn_proto),
+        );
+    ob_proto_write
+        .as_script_object_mut()
+        .unwrap()
+        .force_set_function(
+            "isPropertyEnumerable",
+            is_property_enumerable,
+            gc_context,
+            DontDelete | DontEnum,
+            Some(fn_proto),
+        );
+    ob_proto_write
+        .as_script_object_mut()
+        .unwrap()
+        .force_set_function(
+            "isPrototypeOf",
+            is_prototype_of,
+            gc_context,
+            DontDelete | DontEnum,
+            Some(fn_proto),
+        );
+    ob_proto_write
+        .as_script_object_mut()
+        .unwrap()
+        .force_set_function(
+            "toString",
+            to_string,
+            gc_context,
+            DontDelete | DontEnum,
+            Some(fn_proto),
+        );
 }
