+++ core/src/player.rs
@@ -1,4 +1,4 @@
-use crate::avm1::{self, Avm1};
+use crate::avm1::{Avm1, Value};
 use crate::backend::{
     audio::AudioBackend, navigator::NavigatorBackend, render::Letterbox, render::RenderBackend,
 };
@@ -337,7 +337,7 @@ impl<Audio: AudioBackend, Renderer: RenderBackend, Navigator: NavigatorBackend>
                 start_clip: gc_root.root,
                 target_clip: Some(gc_root.root),
                 root: gc_root.root,
-                target_path: avm1::Value::Undefined,
+                target_path: Value::Undefined,
                 system_prototypes: gc_root.avm.read().prototypes().clone(),
             };
 
@@ -416,7 +416,7 @@ impl<Audio: AudioBackend, Renderer: RenderBackend, Navigator: NavigatorBackend>
                     start_clip: gc_root.root,
                     target_clip: Some(gc_root.root),
                     root: gc_root.root,
-                    target_path: avm1::Value::Undefined,
+                    target_path: Value::Undefined,
                     system_prototypes: gc_root.avm.read().prototypes().clone(),
                 };
 
@@ -487,7 +487,7 @@ impl<Audio: AudioBackend, Renderer: RenderBackend, Navigator: NavigatorBackend>
                 start_clip: gc_root.root,
                 target_clip: Some(gc_root.root),
                 root: gc_root.root,
-                target_path: avm1::Value::Undefined,
+                target_path: Value::Undefined,
                 system_prototypes: gc_root.avm.read().prototypes().clone(),
             };
 
@@ -551,7 +551,7 @@ impl<Audio: AudioBackend, Renderer: RenderBackend, Navigator: NavigatorBackend>
                 start_clip: gc_root.root,
                 target_clip: Some(gc_root.root),
                 root: gc_root.root,
-                target_path: avm1::Value::Undefined,
+                target_path: Value::Undefined,
                 system_prototypes: gc_root.avm.read().prototypes().clone(),
             };
